# swagger/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "AppController_healthCheck",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "App"
        ]
      }
    },
    "/ordenador": {
      "get": {
        "operationId": "OrdenadorController_getOrdenadores",
        "parameters": [
          {
            "name": "rol",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ordenadores encontrados exitosamente"
          },
          "400": {
            "description": "Solicitud incorrecta"
          },
          "404": {
            "description": "Ordenadores no encontrados"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "summary": "Consulta ordenadores por cargo",
        "tags": [
          "Ordenador"
        ]
      }
    },
    "/supervisor/dependencia": {
      "get": {
        "operationId": "SupervisorController_getSupervisorDependencia",
        "parameters": [
          {
            "name": "dependencia",
            "required": true,
            "in": "query",
            "description": "ID de la dependencia",
            "schema": {
              "example": "DEP123",
              "type": "string"
            }
          },
          {
            "name": "fecha",
            "required": true,
            "in": "query",
            "description": "Fecha de consulta (YYYY-MM-DD)",
            "schema": {
              "example": "2024-12-29",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Supervisores encontrados exitosamente"
          },
          "404": {
            "description": "No se encontraron supervisores"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "summary": "Obtener supervisores por dependencia",
        "tags": [
          "Supervisores"
        ]
      }
    },
    "/supervisor/documento": {
      "get": {
        "operationId": "SupervisorController_getSupervisorPorDocumento",
        "parameters": [
          {
            "name": "documento",
            "required": true,
            "in": "query",
            "description": "Número de documento del supervisor",
            "schema": {
              "example": "123456789",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Franjas del supervisor encontrados exitosamente"
          },
          "400": {
            "description": "Formato de documento inválido"
          },
          "404": {
            "description": "No se encontraron Franjas para el documento especificado"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "summary": "Obtener supervisores por número de documento",
        "tags": [
          "Supervisores"
        ]
      }
    },
    "/rol-ordenador": {
      "get": {
        "operationId": "RolOrdenadorController_getOrdenadores",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Consulta roles de ordenadores",
        "tags": [
          "RolOrdenador"
        ]
      }
    },
    "/rol-ordenador/actual": {
      "get": {
        "operationId": "RolOrdenadorController_getOrdenadorActual",
        "parameters": [
          {
            "name": "rol",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ordenador encontrado exitosamente"
          },
          "400": {
            "description": "Solicitud incorrecta"
          },
          "404": {
            "description": "Ordenadores no encontrados"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "summary": "Consulta el ordenador actual por cargo",
        "tags": [
          "RolOrdenador"
        ]
      }
    }
  },
  "info": {
    "title": "Ordenadores Supervisores API MID",
    "description": "API",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [
    {
      "name": "API",
      "description": ""
    }
  ],
  "servers": [],
  "components": {
    "schemas": {}
  }
}